requirements.txt
solidpython
mypy
